Moved default Language to General
