+++ src/router/index.js
@@ -28,8 +28,6 @@ const BomFormats = () =>
   import('@/views/administration/configuration/BomFormats');
 const WelcomeMessage = () =>
   import('@/views/administration/configuration/WelcomeMessage');
-const DefaultLanguage = () =>
-  import('@/views/administration/configuration/DefaultLanguage');
 const Email = () => import('@/views/administration/configuration/Email');
 const Jira = () => import('@/views/administration/configuration/JiraConfig');
 const InternalComponents = () =>
@@ -355,16 +353,6 @@ function configRoutes() {
             {
               path: 'configuration/welcomeMessage',
               component: WelcomeMessage,
-              meta: {
-                title: i18n.t('message.administration'),
-                i18n: 'message.administration',
-                sectionPath: '/admin',
-                permission: 'SYSTEM_CONFIGURATION',
-              },
-            },
-            {
-              path: 'configuration/defaultLanguage',
-              component: DefaultLanguage,
               meta: {
                 title: i18n.t('message.administration'),
                 i18n: 'message.administration',
